fix: extract card styles + typed STYLE_SECTIONS + audit hardening
Move ~2,085 LOC of CSS-in-JS from src/index.js into src/styles.ts; move
the editor's STYLE_SECTIONS array into a discriminated-union typed module
at src/config/styling-config.ts that the editor still consumes unchanged.
Extract _injectLiveFillStyle to src/ui/live-fill-style.ts as the first
inhabitant of src/ui/ (PR 12 will populate the rest). Single source of
truth for radius defaults via genStyleSectionDefaults() emitted into :host.

User-visible fixes
- bar_opacity slider reset moved a fresh card from 30 → 45; STYLE_SECTIONS
  default now tracks DEFAULT_BAR_OPACITY (30) so the reset behaves.
- "Background color" / "Bar background" / "Background" labels in the
  styling tab normalized to consistent "Background" / "Bar background".

Type wins
- Discriminated union (ColorControl | RadiusControl | SliderControl |
  SelectControl) replaces the loose JS shape. Adding a control without
  its required fields now fails tsc.
- CssVarKey template-literal type enforced on every variable. Caught --r
  living outside the --cgc-* namespace (renamed --r → --cgc-card-radius).
- Radius / slider ranges sourced from const.ts so the struct validator
  and editor share one set of bounds (no more drift like A2).

CSS cleanup during extraction
- Removed 13 dead CSS classes never rendered in JS (.tsleft, .tspill,
  .tspill-left, .bulkbar-check, .empty.inpreview, .filter-empty,
  .live-pill-btn.mic-error, .cgc-debug-val.status-{ok,warn,bad},
  .tsbar-{back,cam}-btn, .tsbar-live-center)
- Deduped --cgc-pill-bg in :host (one was dead, #000 always winning)
- Dropped dead "position: relative;" in .live-picker (overridden on next
  line by "position: absolute;")
- Renamed inconsistent internal vars to the public --cgc-* namespace:
  --barOpacity, --thumbRowH, --thumbEmptyH, --thumbsMaxHeight, --topbarMar,
  --topbarPad, --tgap → --cgc-bar-opacity, --cgc-thumb-row-h, etc.
- Dropped --gap (set but never read)

CSS compaction
- Native CSS nesting (Chrome 120+, Safari 16.5+, Firefox 117+) collapsed
  14 selector families: :host([data-live-fs]), .live-card-host,
  .live-picker, .gallery-pills, .gallery-pill, .controls-bar-fixed,
  .tthumb, .tthumbs, .tph, .tbar, .objbtn, .segbtn, .thumb-menu-item,
  .bulkaction, .cgc-debug-*
- :is() consolidations replaced repeated multi-selector groups, e.g.
  .tthumb.bar-{top,bottom}.with-bar :is(.tph.broken, .tph.disabled,
  .tph.spinner) — also surfaced an exact 5-rule duplicate between
  .tph.broken ha-icon and .tph.disabled ha-icon
- Shimmer animation declarations deduped 3× → 1× across .tph.skeleton,
  .preview-skeleton, .live-card-host:empty
- Two identical prefers-reduced-motion blocks merged
- Net: src/styles.ts is 1,942 LOC vs the 2,138 a verbatim move would
  have produced (~200 LOC removed via smarter CSS)

Specs
- src/config/styling-config.spec.ts (24 tests): structural invariants
  (unique ids/variables/configKeys, --cgc-* namespace, min<max,
  default ∈ [min,max], unique select options), STYLE_SECTIONS ↔ const.ts
  cross-checks (regression guard for A2), genStyleSectionDefaults output.
- src/ui/live-fill-style.spec.ts (8 tests, jsdom env): idempotency,
  nested shadow-root walking, padding-bottom neutralization, 2s reinject,
  MutationObserver fallback.
- jsdom added as devDep.

Out of scope (PR 11/12 territory)
- Editor rewrite to ha-form schemas — STYLE_SECTIONS types are designed
  to be consumed unchanged.
- UI sub-component split — src/ui/ created with one inhabitant for now.
- z-index magic numbers (14 distinct values, 1 known collision at
  z-index:30 between .cgc-debug-modal and .bulk-floating-hint) and
  !important audit (29 occurrences) — defer to a dedicated cleanup PR.

Bundle: 380 KB (flat vs main).
src/index.js: 12,029 → 9,799 lines (−2,230).
