+++ package.json
@@ -54,6 +54,7 @@
     "@rollup/plugin-replace": "^6.0.1",
     "@rollup/plugin-terser": "^1.0.0",
     "@rollup/plugin-typescript": "^12.1.1",
+    "@vitest/coverage-v8": "^4.1.5",
     "chokidar": "^5.0.0",
     "concurrently": "^9.0.1",
     "dotenv": "^17.4.2",
@@ -62,13 +63,13 @@
     "eslint-plugin-wc": "^3.1.0",
     "globals": "^17.6.0",
     "husky": "^9.1.6",
+    "jsdom": "^29.1.1",
     "lint-staged": "^17.0.4",
     "prettier": "^3.3.3",
     "rollup": "^4.60.3",
     "tslib": "^2.8.0",
     "typescript": "^6.0.3",
     "typescript-eslint": "^8.59.2",
-    "vitest": "^4.1.5",
-    "@vitest/coverage-v8": "^4.1.5"
+    "vitest": "^4.1.5"
   }
 }
